A few remaining bugs
In addition to the issue with editing an existing todo items text,
there is an issue with routing. The window.location gets parsed
and a route action gets dispatched.  However, my subscription to
the store doesn't seem to be getting called as a result.  So either
there is something wrong with the subscription or the dispatched
action simply isn't getting processed.  I'm not sure which yet.